+++ Ponkan/App/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "clementine(1).png",
+      "filename" : "logo.png",
       "idiom" : "universal",
       "platform" : "ios",
       "size" : "1024x1024"
