+++ Ponkan/Interface/Sugar/Sources/Sugar/Text+.swift
@@ -2,8 +2,8 @@ import SwiftUI
 
 extension Text {
     public init(_ string: String, configure: ((inout AttributedString) -> Void)) {
-        var attributedString = AttributedString(string) /// create an `AttributedString`
-        configure(&attributedString) /// configure using the closure
-        self.init(attributedString) /// initialize a `Text`
+        var attributedString = AttributedString(string)
+        configure(&attributedString)
+        self.init(attributedString)
     }
 }
